// frontend/src/app/core/models/admin-dashboard.ts
export interface AdminDashboardResponse {

  properties: number;

  users: number;

  clients: number;

  agents: number;

  visits: number;

  appointments: number;

  transactions: number;

  payments: number;

  pendingVisits: number;

  pendingAppointments: number;

  activeTransactions: number;

  totalPayments: number;
}